Adding keys to list items
Key are like an ID that helps keep everything in sync.
Adjust our dishes array to instead of returning strings,
we can return objects. We need to write a little
transformation function, to map over the dishes.
For each dish, we're going to return an object
and whenever we return an object from a function, in line,
we need to wrap it in parenthesis.