Destructuring arrays and objects
Access array and assign it to a variable, index.js
Create placeholders for unwanted array elements with `,`
Access props object can help us when we have a lot of property names.
Object destructuring is going to make the syntax in our components
a little cleaner, by destructuring the props object, and then
array destructuring is going to help us in a second,
when we start to talk about state.