Understanding the useState Hook
useState returns an array of two elements.
The first item in the array is our state variable.
The second item is a function to update state.
Array destructuring is used to pop out the first value,
give it a name and set the initial value of the state variable.
This allows us to manage any sort of local variables
we want to keep track of.