Working with useEffect
Used to manage side effects that aren't related to
the components render. So things like console messages,
loading data, and sometimes animations can benefit from useEffect.
useEffect's second argument is called the dependency array.
If you pass an empty dependency array, the props and state
inside the effect will always have the initial values.
So the effect is not going to be called again after the
first render. You can use this array to keep track of values so
useEffect will pass `emotion` every time this is being updated.