Incorporating useReducer
A reducer function's most simple definition is
it takes in the current state and it returns a new state.
useReducer takes in two arguments. The first argument
is the function. The second argument is the same as
the first argument in useState, it's the initial state.
A little upgrade from useState to useReducer. A reducer function
takes in the current state and returns a new state. Then,
we can use that function to update the state for the checkbox.