Fetching data with Hooks
useEffect will grab data from the GitHub API.
useState will manage the data.
Pass useEffect an empty array, because we only want
to fetch data when the component renders.
If there's some data, we want to return a div with the data.
Otherwise a message in a div, explaining the problem.
Uses props destructure.