Testing small functions with Jest
Automatically included when we use create-react-app
Import the file under test.
`test` takes two arguments.
The first argument to is the name of the test, "multiplies by two".
The second argument is a callback function and
contains the assertion to call the function with required arguments,
i.e. expect and matcher.