Introducing React testing library
Automatically included when we use create-react-app
Tests if rendering elements is correct
In test file, import "@testing-library/react" and "React"
In this case, use `getByText` to destructure `render`,
which is called a query.
With React Testing Library, a query is going to reach out
to some sort of element and return some information about it.