Testing Hooks with React testing library
Revisit the checkbox component which used a hook called useReducer.
Deconstruct render of component using the query `getByLabelText` selector.
Selected the component by the initial state when loading the component.
`fireEvent` from testing library to handle the `onClick` event in the component.
Expectation as the last step in the testing file.