Configuring the router
The router is going to live in this index JS file.
To pass all of the information from the router to any nested components.
`react-router-dom` is because we're using the DOM, we're using a website,
but they also have tools for native applications as well.
Wrap our app component in the router, this'll give the app access to all
the properties of the router, i.e. the location, the history, etc,
which lives in the router.
`App.js` is responsible telling the router which component to render
whenever the location changes.
May need to `npm install history`
`yarn start`, go to `localhost:3000/events` for the events page, etc