Incorporating the link component
In pages.js adjust it to use a Link component,
from react-router-dom, `<Link to="about">About</Link>`.
inside a `<nav>`. This property links to a particular page.

Handling incorrect page routing with `Whoops404`. This is
added to our route config in our App.js file. Replace the path
with a star or an asterisk, which creates a wild card catch all.

One step even better, is to use from react-router-dom, a hook
called useLocation. It is a function to return our current
page location object, with a few different keys. `location.pathname`
can be added to error message using the useLocation hook.